# src/openapi.yaml
openapi: 3.0.3
info:
  title: ScoutOff API
  description: >
    Backend API for ScoutOff — a decentralised football scouting platform built on the Stellar
    blockchain (Soroban smart contracts).
  version: 1.0.0
  contact:
    name: ScoutOff Engineering
servers:
  - url: /api
    description: Stable API (version-less alias)
  - url: /api/v1
    description: API v1
  - url: /api/v2
    description: API v2 (includes SSE stream endpoint)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Long-lived scout API key (alternative to a Bearer JWT). Restricted keys may only perform
        operations covered by their granted scopes (#1019). Legacy keys without a scope list are
        unrestricted.
...
paths:
...
  /admin/validators:
    get:
      tags:
        - admin
      summary: Returns the full list of registered validator wallets.
      operationId: listValidators
      description: |-
        Returns the full list of registered validator wallets from the local DB,
        including their registration timestamp, revocation timestamp (if any), and tx_hash.
      responses:
        '200':
          description: '{ success: true, data: ValidatorRow[] }'
      security:
        - bearerAuth: []
      x-required-role:
        - admin
...
  /validators/milestone:
    post:
      tags:
        - validators
      summary: Submit evidence for a player milestone.
      operationId: submitMilestoneEvidence
      description: |-
        Submit evidence for a player milestone. `evidenceUri` may be an
        `https://` URL (downloaded and re-pinned to IPFS) or an `ipfs://` CID
        (recorded directly). Invalidates the player's milestone cache on success.

        Rate-limited per IP, per validator wallet, and per target player_id (#1137).
      requestBody:
        required: true
        description: >-
          { playerId: string, milestoneType: 'identity'|'performance'|'trial_offer', evidenceUri:
          string }
        content:
          application/json:
            schema:
              type: object
              properties:
                playerId:
                  type: string
                  minLength: 1
                milestoneType:
                  type: string
                  enum:
                    - identity
                    - performance
                    - trial_offer
                evidenceUri:
                  type: string
                  minLength: 1
                  description: Additional validation is applied beyond this schema — see the route handler.
                notes:
                  type: string
                  maxLength: 2000
                validatorComment:
                  type: string
                  maxLength: 2000
              required:
                - playerId
                - milestoneType
                - evidenceUri
              additionalProperties: false
      responses:
        '201':
          description: '{ success: true, data: { evidenceCid: string } }'
        '400':
          description: '{ success: false, error: string } - Invalid body'
        '413':
          description: '{ success: false, error: string } - Remote evidence file too large'
        '422':
          description: '{ success: false, error: string } - Remote evidence has an unsupported content type'
        '429':
          description: >-
            { success: false, error: string } - Per-player rate limit exceeded; Retry-After header
            set
      security:
        - bearerAuth: []
      x-required-role:
        - validator
  /validators/milestones/approve-bulk:
    post:
      tags:
        - validators
      summary: Approve a batch of milestones assigned to the authenticated validator in one call.
      operationId: approveBulkMilestones
      description: |-
        Approve a batch of milestones assigned to the authenticated validator in
        one call. Each ID is processed independently — a failure on one does not
        abort the batch; per-ID outcomes are returned in `data`.
      requestBody:
        required: true
        description: '{ milestoneIds: string[] } - At least one ID required'
        content:
          application/json:
            schema:
              type: object
              properties:
                milestoneIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
              required:
                - milestoneIds
              additionalProperties: false
      responses:
        '200':
          description: >-
            { success: true, data: Array<{ milestoneId, status:
            'approved'|'invalid'|'unauthorized'|'error', error? }> }
        '400':
          description: '{ success: false, error: string } - Empty milestoneIds'
      security:
        - bearerAuth: []
      x-required-role:
        - validator
...
  /validators/{wallet}/stats:
    get:
      tags:
        - validators
      summary: >-
        Dashboard summary for a validator: pending count, all-time approved/rejected totals,
        approvals in the last 30 days, and up to 20 recent a...
      operationId: getValidatorDashboardStats
      description: |-
        Dashboard summary for a validator: pending count, all-time approved/rejected
        totals, approvals in the last 30 days, and up to 20 recent activity events.

        Counts are derived from indexed milestone events — no separate write path.
      parameters:
        - name: wallet
          in: path
          required: true
          schema:
            type: string
          description: Validator's Stellar public key
      responses:
        '200':
          description: >-
            { success: true, data: { wallet, pending, approvedTotal, rejectedTotal, approvedLast30d,
            recent: [...] } }
        '401':
          description: '{ success: false, error: string } - Missing or invalid token'
        '403':
          description: '{ success: false, error: string } - Insufficient permissions or wallet mismatch'
      security:
        - bearerAuth: []
      x-required-role:
        - validator
      x-owner-only: true
...